Reorganization of low level library
