+++ Host/API/device_struct_handler.c
